fix/deleteFolder

+++ src/main/java/ecc/sogonsogon/BE/service/FolderService.java
@@ -29,8 +29,8 @@ public class FolderService {
         //1. 대상 찾기
         Folder target = folderRepository.findById(id).orElse(null);
         //2. 잘못된 요청 처리
-        if(target == null) {
-            return null;
+        if(target == null || !target.getUser().equals(user)) {
+            throw new IllegalArgumentException("삭제할 수 없습니다.");
         }
         //3. 대상 삭제
         folderRepository.delete(target);
@@ -41,8 +41,8 @@ public class FolderService {
 
         Folder target=folderRepository.findById(id).orElse(null);
         //3. 잘못된 요청 처리
-        if(target == null) {
-            return null;
+        if(target == null || !target.getUser().equals(user)) {
+            throw new IllegalArgumentException("폴더 이름을 수정할 수 없습니다.");
         }
         //4. 업데이트 및 정상 응답
         if (folderDto.getFolderName() != null) {
